Make `generate_index_type` into a macro-by-example
