+++ charon/macros/src/lib.rs
@@ -4,7 +4,7 @@
 
 extern crate proc_macro;
 extern crate syn;
-use proc_macro::{Span, TokenStream, TokenTree};
+use proc_macro::{Span, TokenStream};
 use quote::ToTokens;
 use serde::Deserialize;
 use std::fs::File;
@@ -22,85 +22,6 @@ use syn::{
 const _TAB: &'static str = "    ";
 const _TWO_TABS: &'static str = "        ";
 const THREE_TABS: &'static str = "            ";
-
-/// This is very annoying, but we can't use a global constant string in `format`:
-/// we need to define a macro to return a string literal.
-/// TODO: turn [generic_index_type] into a simpl macro rule.
-macro_rules! index_generic_code {
-    () => {
-        "
-pub mod {} {{
-    index_vec::define_index_type! {{
-        pub struct Id = usize;
-        // Must fit in an u32 for serialization.
-        MAX_INDEX = std::u32::MAX as usize;
-    }}
-
-    pub type Vector<T> = crate::id_vector::Vector<Id, T>;
-    pub type Map<T> = crate::id_map::Map<Id,T>;
-    pub type Generator = crate::id_generator::Generator<Id>;
-    pub type MapGenerator<K> = crate::id_generator::MapGenerator<K, Id>;
-
-    impl Id {{
-        pub fn is_zero(&self) -> bool {{
-            self.index() == 0
-        }}
-
-        pub fn incr(&mut self) {{
-            // Overflows are extremely unlikely, but we do want to make sure
-            // we panick whenever there is one.
-            *self = Self::new(self.index().checked_add(1).unwrap());
-        }}
-    }}
-
-    pub static ZERO: Id = Id {{ _raw: 0 }};
-    pub static ONE: Id = Id {{ _raw: 1 }};
-
-    impl std::fmt::Display for Id {{
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) ->
-          std::result::Result<(), std::fmt::Error> {{
-            f.write_str(self.index().to_string().as_str())
-        }}
-    }}
-}}"
-    };
-}
-
-/// Generate an `Index` module which contains an index type and a generator
-/// for fresh indices. We use it because we need manipulate
-/// a lot of different indices (for various kinds of declarations, variables, blocks,
-/// etc.).
-/// For sanity purposes, we prevent any confusion between the different kinds
-/// of indices by using different types. The following macro allows us to
-/// easily derive those types, and the needed utilities (casts, vectors indexed
-/// by those opaque indices, etc.).
-///
-/// The `ident` parameter should contain the name of the module to declare.
-#[proc_macro]
-pub fn generate_index_type(item: TokenStream) -> TokenStream {
-    // Check that the token strem is made of exactly one identifier
-    let mut tokens = vec![];
-    for token in item {
-        tokens.push(token);
-    }
-    if tokens.len() != 1 {
-        panic!("generate_index_type: invalid parameters: should receive exactly one identifier");
-    }
-    let token = &tokens[0];
-    match token {
-        TokenTree::Ident(ident) => {
-            // Generate the index code
-            let code: String = format!(index_generic_code!(), ident.to_string());
-            let code: TokenStream = code.parse().unwrap();
-            return code;
-        }
-        _ => {
-            panic!(
-                "generate_index_type: invalid parameters: should receive exactly one identifier"
-            );
-        }
-    }
-}
 
 macro_rules! derive_variant_name_impl_code {
     () => {
